+++ test/fuzz/_common.cjs
@@ -62,9 +62,9 @@ function getExpectedOutput({ arg, shell }, normalizeWhitespace) {
       arg = arg.replace(/[\t\n\r ]+/gu, " ");
     } else {
       arg = arg
-        .replace(/[\t\n ]+/gu, " ")
+        .replace(/\n/gu, " ")
         .replace(/\r(?!\n)/gu, "")
-        .replace(/[\t\n ]+/gu, " ");
+        .replace(/\n/gu, " ");
     }
 
     // Trim the string, like the shell
@@ -72,8 +72,6 @@ function getExpectedOutput({ arg, shell }, normalizeWhitespace) {
       arg = arg.replace(/^[\s\u0085]+/gu, "");
     } else if (isShellCmd(shell)) {
       arg = arg.replace(/^[\t\n\r ]+|(?<![\t\n\r ])[\t\n\r ]+$/gu, "");
-    } else {
-      arg = arg.replace(/^[\t\n ]+|(?<![\t\n ])[\t\n ]+$/gu, "");
     }
   } else {
     // Change newlines to spaces, like Shescape